improve SQL injection

+++ admin/includes/view_all_users.php
@@ -56,6 +56,7 @@ if (isset($_GET['change_to_admin'])) {
     $query                 = "UPDATE users SET user_role = 'admin' WHERE user_id = $the_user_id";
     $make_admin_user_query = mysqli_query($connection, $query);
     confirm_query($make_admin_user_query);
+    $_SESSION['user_role'] = 'admin';
     header("Location: users.php");
 }
 
@@ -65,15 +66,25 @@ if (isset($_GET['change_to_sub'])) {
     $query               = "UPDATE users SET user_role = 'subscriber' WHERE user_id = $the_user_id";
     $make_sub_user_query = mysqli_query($connection, $query);
     confirm_query($make_sub_user_query);
+    $_SESSION['user_role'] = 'subscriber';
     header("Location: users.php");
 }
 
 
 if (isset($_GET['delete'])) {
-    $the_user_id       = $_GET['delete'];
-    $query             = "DELETE FROM users WHERE user_id = $the_user_id";
-    $delete_user_query = mysqli_query($connection, $query);
-    confirm_query($delete_user_query);
-    header("Location: users.php");
+    if (isset($_SESSION['user_role'])) {
+        if ($_SESSION['user_role'] == 'admin') {
+            $the_user_id       = mysqli_real_escape_string($connection, $_GET['delete']);
+            $query             = "DELETE FROM users WHERE user_id = $the_user_id";
+            $delete_user_query = mysqli_query($connection, $query);
+            confirm_query($delete_user_query);
+            header("Location: users.php");
+        } else {
+            echo "YOU are not an Admin!";
+        }
+    } else {
+        echo "Not working";
+    }
+
 }
 ?>
